pkg/sandbox: add InitInfraContainer endpoint
Eventually, we will move all of the infra container configuration into this package
